完善Spring Security和JWT的安全认证机制

+++ src/main/java/com/cgs/gaexam/web/controller/RoleController.java
@@ -27,7 +27,7 @@ public class RoleController {
     }
 
     @DeleteMapping("/{id}")
-    public Result delete(@PathVariable Integer id) {
+    public Result delete(@PathVariable Long id) {
         roleService.deleteById(id);
         return ResultGenerator.genSuccessResult();
     }
@@ -39,7 +39,7 @@ public class RoleController {
     }
 
     @GetMapping("/{id}")
-    public Result detail(@PathVariable Integer id) {
+    public Result detail(@PathVariable Long id) {
         Role role = roleService.findById(id);
         return ResultGenerator.genSuccessResult(role);
     }
